Auto-adjust slowpoll wait time (and remove manual buffer setting)
The manual buffer did not (always) work as intended. Probably there was too
little CPU left to catch up with processing the packets.

Instead, we'll auto-adjust the time that libslowpoll will wait. By reducing the
wait-time based on the amount of packets that never reached us, we can
properly sniff all packets.

This should fix the interesting behaviour where:

- when 80K packets per second were flowing
- if just the 'udp' filter was
- only some streams would get reported on
- but if more a specific network was filtered, more info was shown

That was due to the buffers being too full.